requirements.txt
vobject
pandas
numpy
scikit-learn
tqdm
fuzzywuzzy
phonenumbers
PyYAML
requests
openai
python-dotenv
